+++ src/components/cubes/Modal.tsx
@@ -1,96 +1,25 @@
 import InputText from "../input-text/InputText";
 import CheckboxImage from "./CheckboxImage";
-import { Categories } from "@/interfaces/Categories";
 import { cubeCollection } from "@/lib/cubeCollection";
 import genId from "@/lib/genId";
-import { useCubesModalStore } from "@/store/CubesModalStore";
-import createCube from "@/lib/createCube";
-import { useTimerStore } from "@/store/timerStore";
-import loadCubes from "@/lib/loadCubes";
 import translation from "@/translations/global.json";
-import { useSettingsModalStore } from "@/store/SettingsModalStore";
-import { useState } from "react";
+import useModalCube from "@/hooks/useModalCube";
+import { useCubesModalStore } from "@/store/CubesModalStore";
 
-export default function ModalCreate() {
+export default function Modal() {
+  const { editingCube } = useCubesModalStore();
   const {
-    setModalOpen,
-    editingCube,
-    setEditingCube,
+    error,
+    handleClickRadio,
+    handleWriteCubeName,
+    handleCreateCube,
+    handleEditCube,
+    handleDeleteCube,
+    handleCloseModal,
     selectedCategory,
-    setSelectedCategory,
     cubeName,
-    setCubeName,
-  } = useCubesModalStore();
-  const { lang } = useSettingsModalStore();
-  const { setCubes } = useTimerStore();
-  const [error, setError] = useState<boolean>(false);
-
-  const handleClickRadio = (category: Categories) => {
-    setSelectedCategory(category);
-  };
-
-  const handleWriteCubeName = (newText: string) => {
-    if (newText.trim().length > 0) setError(false);
-    setCubeName(newText);
-  };
-
-  const handleCreateCube = (name: string, category: Categories) => {
-    if (name.trim() === "") {
-      setError(true);
-      return;
-    }
-    const newCubes = createCube({
-      cubeName: name,
-      category: category,
-    });
-    setCubes(newCubes);
-    setModalOpen(false);
-    setEditingCube(null);
-    setCubeName("");
-    setSelectedCategory("2x2");
-  };
-
-  const handleEditCube = (name: string, category: Categories) => {
-    if (name.trim() === "") {
-      setError(true);
-      return;
-    }
-    if (!editingCube) return;
-    const cubeDB = loadCubes();
-
-    for (const cube of cubeDB) {
-      if (cube.id === editingCube.id) {
-        cube.name = name;
-        cube.category = category;
-      }
-    }
-
-    window.localStorage.setItem("cubes", JSON.stringify(cubeDB));
-    setCubes(cubeDB);
-    setModalOpen(false);
-    setEditingCube(null);
-    setCubeName("");
-    setSelectedCategory("2x2");
-  };
-
-  const handleDeleteCube = () => {
-    const cubeDB = loadCubes();
-    if (!editingCube) return;
-    const updatedCubeDB = cubeDB.filter((cube) => cube.id !== editingCube.id);
-    window.localStorage.setItem("cubes", JSON.stringify(updatedCubeDB));
-    setCubes(updatedCubeDB);
-    setModalOpen(false);
-    setEditingCube(null);
-    setCubeName("");
-    setSelectedCategory("2x2");
-  };
-
-  const handleCloseModal = () => {
-    setModalOpen(false);
-    setEditingCube(null);
-    setCubeName("");
-    setSelectedCategory("2x2");
-  };
+    lang,
+  } = useModalCube();
 
   return (
     <>
@@ -101,7 +30,7 @@ export default function ModalCreate() {
         aria-hidden={false}
         className={`fixed backdrop-blur-[2px] top-0 left-0 z-50 w-full p-4 overflow-x-hidden overflow-y-auto md:inset-0 h-screen flex flex-col items-center`}
         onClick={(e) => {
-          if (e.target === e.currentTarget) setModalOpen(false);
+          if (e.target === e.currentTarget) handleCloseModal();
         }}
       >
         <div className="relative w-full max-w-2xl max-h-full">
