+++ src/components/cubes/OptionsContainer.tsx
@@ -1,0 +1,11 @@
+interface OptionsContainer {
+  children: React.ReactNode;
+}
+
+export function OptionsContainer({ children }: OptionsContainer) {
+  return (
+    <>
+      <div className="flex justify-end gap-3">{children}</div>
+    </>
+  );
+}
